+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-elements-writer
 
+## 1.3.37
+
+### Patch Changes
+
+-   Updated dependencies [036b48b]
+    -   @sap-ux/odata-service-writer@0.25.0
+
+## 1.3.36
+
+### Patch Changes
+
+-   Updated dependencies [9f14de1]
+    -   @sap-ux/odata-service-writer@0.24.2
+
 ## 1.3.35
 
 ### Patch Changes